// frontend/src/app/interfaces/source.ts
export interface Source {
  id: string;
  name: string;
  url: string;
  isFamilyFriendly: boolean;
  displayUrl: string;
  snippet: string;
  deepLinks: { snippet: string; name: string; url: string }[];
  dateLastCrawled: string;
  cachedPageUrl: string;
  language: string;
  primaryImageOfPage?: {
    thumbnailUrl: string;
    width: number;
    height: number;
    imageId: string;
  };
  attn: {
    name: number[];
    snippet: number[];
    score: number;
  }[];
  isNavigational: boolean;
  nameTokens: string[];
  snippetTokens: string[];
}